+++ src/Information/InformationMeasures.jl
@@ -1,6 +1,6 @@
-abstract InformationMeasure{T}
+abstract AbstractMeasure{T}
 
-abstract SymmetricMeasure{T} <: InformationMeasure{T}
+abstract SymmetricMeasure{T} <: AbstractMeasure{T}
 
 # Entropy
 # =======
@@ -162,9 +162,9 @@ immutable GapUnionPercentage{T} <: SymmetricMeasure{T} end
 immutable GapIntersectionPercentage{T} <: SymmetricMeasure{T} end
 
 function estimate{B, T}(measure::GapIntersectionPercentage{B}, nxy::ResidueCount{T, 2, true})
-  B(nxy[21, 21]) / B(sum(nxy))
+  100.0 * B(nxy[21, 21]) / B(sum(nxy))
 end
 
 function estimate{B, T}(measure::GapUnionPercentage{B}, nxy::ResidueCount{T, 2, true})
-  B(nxy.marginals[21, 1] + nxy.marginals[21, 2] - nxy[21, 21]) / B(sum(nxy))
+  100.0 * B(nxy.marginals[21, 1] + nxy.marginals[21, 2] - nxy[21, 21]) / B(sum(nxy))
 end