+++ test/buslje09.jl
@@ -376,12 +376,12 @@ let file = joinpath(pwd(), "data", "gaps.txt")
   @test_approx_eq gu[1, 1] 0.0
   @test_approx_eq gi[1, 1] 0.0
 
-  @test_approx_eq gu[1, 2] getweight(cl, 10)/ncl
+  @test_approx_eq gu[1, 2] 100.0 * getweight(cl, 10)/ncl
   @test_approx_eq gi[1, 2] 0.0
 
-  @test_approx_eq gu[10, 9] (ncl - getweight(cl, 1))/ncl
-  @test_approx_eq gi[10, 9] (ncl - getweight(cl, 1) - getweight(cl, 2))/ncl
+  @test_approx_eq gu[10, 9] 100.0 * (ncl - getweight(cl, 1))/ncl
+  @test_approx_eq gi[10, 9] 100.0 * (ncl - getweight(cl, 1) - getweight(cl, 2))/ncl
 
-  @test_approx_eq gu[10, 10] (ncl - getweight(cl, 1))/ncl
-  @test_approx_eq gu[10, 10] (ncl - getweight(cl, 1))/ncl
+  @test_approx_eq gu[10, 10] 100.0 * (ncl - getweight(cl, 1))/ncl
+  @test_approx_eq gu[10, 10] 100.0 * (ncl - getweight(cl, 1))/ncl
 end